refactor setup_py -> package_dists and move most rules to util_rules
