+++ src/python/pants/backend/python/typecheck/mypy/mypyc.py
@@ -5,7 +5,6 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 
-from pants.backend.python.goals.setup_py import DistBuildEnvironment, DistBuildEnvironmentRequest
 from pants.backend.python.target_types import PythonDistribution
 from pants.backend.python.typecheck.mypy.subsystem import (
     MyPy,
@@ -13,6 +12,10 @@ from pants.backend.python.typecheck.mypy.subsystem import (
     MyPyFirstPartyPlugins,
 )
 from pants.backend.python.util_rules import pex_from_targets
+from pants.backend.python.util_rules.package_dists import (
+    DistBuildEnvironment,
+    DistBuildEnvironmentRequest,
+)
 from pants.backend.python.util_rules.pex import Pex, PexRequest
 from pants.backend.python.util_rules.pex_from_targets import RequirementsPexRequest
 from pants.engine.rules import Get, MultiGet, collect_rules, rule
